requirements.txt
torch
torchvision
torchaudio
transformers
accelerate
grpcio-tools
grpcio
protobuf
